Dagster/dagster_requirements.txt
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.5.0
backoff==2.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==14.0
croniter==3.0.3
dagster==1.8.7
dagster-graphql==1.8.7
dagster-pipes==1.8.7
dagster-webserver==1.8.7
docstring_parser==0.16
filelock==3.16.1
fsspec==2024.9.0
gql==3.5.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.1.0
grpcio==1.66.1
grpcio-health-checking==1.62.3
h11==0.14.0
httptools==0.6.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
numpy==2.1.1
packaging==24.1
pandas==2.2.3
protobuf==4.25.5
psutil==6.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyodbc==5.1.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==306
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
structlog==24.4.0
tabulate==0.9.0
tomli==2.0.1
toposort==1.10
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
universal_pathlib==0.2.5
urllib3==2.2.3
uvicorn==0.30.6
watchdog==5.0.2
watchfiles==0.24.0
websockets==13.0.1
yarl==1.11.1
